# project/swagger.yaml
openapi: 3.0.0
info:
  title: Apartment Service API
  version: 1.0.0
servers:
  - url: http://{{baseurl}}
...
paths:
...
  /manager/bill/{apartment-id}:
    post:
      tags:
        - manager > bill > {apartment-id} > create
      summary: Create bill (Manager)
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                bill_type:
                  type: string
                  example: water
                total_amount:
                  type: number
                  example: '99.99'
                due_date:
                  type: string
                  example: '2023-12-31'
                billing_deadline:
                  type: string
                  example: '2023-12-25'
                description:
                  type: string
                  example: Monthly subscription
                bill_image:
                  type: string
                  format: binary
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: >-
            Bearer
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiJmT2xpd0lnQ3dacUYyMUEwOXpnNl9aY0N3dHM3ZHVtM2ctbVFVSFE9IiwidXNlcl90eXBlIjoibWFuYWdlciIsImV4cCI6MTc1NDQwNDE4NywiaWF0IjoxNzU0MzE3Nzg3fQ.fqhaMmROlbK_x3mpfSI30u5Gp-pwsz_GNdM5hxCIMuE
        - name: apartment-id
          in: path
          schema:
            type: integer
          required: true
          example: '1'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
              example: ''
  /manager/bill:
    get:
      tags:
        - manager > bill > get
      summary: Get bill by ID (Manager)
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: >-
            Bearer
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiJmT2xpd0lnQ3dacUYyMUEwOXpnNl9aY0N3dHM3ZHVtM2ctbVFVSFE9IiwidXNlcl90eXBlIjoibWFuYWdlciIsImV4cCI6MTc1NDQwNDE4NywiaWF0IjoxNzU0MzE3Nzg3fQ.fqhaMmROlbK_x3mpfSI30u5Gp-pwsz_GNdM5hxCIMuE
        - name: id
          in: query
          schema:
            type: integer
          example: '1'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
              example: ''
    delete:
      tags:
        - manager > bill > delete
      summary: Delete bill (Manager)
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: >-
            Bearer
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1aWQiOiJmT2xpd0lnQ3dacUYyMUEwOXpnNl9aY0N3dHM3ZHVtM2ctbVFVSFE9IiwidXNlcl90eXBlIjoibWFuYWdlciIsImV4cCI6MTc1NDQwNDE4NywiaWF0IjoxNzU0MzE3Nzg3fQ.fqhaMmROlbK_x3mpfSI30u5Gp-pwsz_GNdM5hxCIMuE
        - name: bill_id
          in: query
          schema:
            type: integer
          example: '1'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
              example: ''
...